✨팩토리 로직 수정
Contents
1.경곗값만 대략적으로 신경쓴 로직을 조금 더 추상화
2.메서드 명 수정

Todo
1.Input에서 1을 받으면 의미 없기 때문에 해당 로직 수정
2.객체를 조금 더 세분화 시키기
