+++ README.md
@@ -14,7 +14,7 @@
 |       | &nbsp; 자체 상태 검증        | &nbsp; 프로젝트 내 대부분의 객체들이 자신의 상태를 관리 ||
 |       | &nbsp; [자체] 기존 오류 수정  | &nbsp; Validation, 로직 수정 ||
 |       | &nbsp; [자체] 클래스 사용    | &nbsp; 사용하지 않던 클래스 연결||
-|       | &nbsp; [자체] 중복 메서드 제거| &nbsp; 중복 메서드 제거||
+|       | &nbsp; [자체] 중복 메서드 제거| &nbsp; 중복 메서드 제거|O|
 
 <br/>
