+++ src/ladder/domain/GameManager.java
@@ -1,8 +1,11 @@
-package ladder.model;
+package ladder.domain;
 
-import ladder.model.ladder.Height;
-import ladder.model.ladder.LadderLines;
+import ladder.domain.ladder.Height;
+import ladder.domain.ladder.LadderLines;
+import ladder.domain.user.Names;
 import ladder.system.Configuration;
+import ladder.utils.Count;
+import ladder.utils.Counts;
 import ladder.view.InputView;
 import ladder.view.OutputView;
 
@@ -28,11 +31,11 @@ public class GameManager {
     }
 
     public void startGame() throws Exception {
+        while(true){
         Names names = inputView.getPlayerNames();
         Height height = inputView.getHeight();
         Count count = Counts.of(names.size());
-
         LadderLines ladderLines = new LadderLines(count, height);
         outputView.print(ladderLines, names);
-    }
+    }}
 }
