+++ src/ladder/domain/LadderGame.java
@@ -1,8 +1,9 @@
-package ladder.model;
+package ladder.domain;
 
-import ladder.model.ladder.Height;
-import ladder.model.ladder.Ladder;
-import ladder.model.ladder.LadderLines;
+import ladder.domain.ladder.Height;
+import ladder.domain.ladder.Ladder;
+import ladder.domain.ladder.LadderLines;
+import ladder.domain.user.Names;
 import ladder.system.Configuration;
 
 public class LadderGame {
