+++ src/ladder/domain/ladder/Height.java
@@ -1,4 +1,4 @@
-package ladder.model.ladder;
+package ladder.domain.ladder;
 
 import java.util.Objects;
