+++ src/ladder/domain/user/Name.java
@@ -1,4 +1,4 @@
-package ladder.model;
+package ladder.domain.user;
 
 public class Name {
